refactor: Remove `trees/` directory. Seems that the AST and CST nodes are very different.
