+++ test/test_node.py
@@ -2,7 +2,7 @@ from unittest import TestCase
 
 from parser.str_region import StrRegion
 from parser.tokens import NumberToken, IdentNameToken
-from parser.trees.named_node import node_from_token
+from parser.cst.named_node import node_from_token
 from parser.cst.nodes import NumberNode, IdentNode
